NP-1500: split api.model to requests and responses
