+++ create-doi-request/src/test/java/no/unit/nva/doi/requests/handlers/CreateDoiRequestHandlerTest.java
@@ -22,8 +22,8 @@ import java.time.ZoneId;
 import java.util.Map;
 import java.util.UUID;
 import no.unit.nva.doi.requests.contants.ServiceConstants;
-import no.unit.nva.doi.requests.api.model.CreateDoiRequest;
-import no.unit.nva.doi.requests.api.model.DoiRequestSummary;
+import no.unit.nva.doi.requests.api.model.requests.CreateDoiRequest;
+import no.unit.nva.doi.requests.api.model.responses.DoiRequestSummary;
 import no.unit.nva.doi.requests.service.DoiRequestsService;
 import no.unit.nva.doi.requests.service.impl.DynamoDBDoiRequestsService;
 import no.unit.nva.doi.requests.util.DoiRequestsDynamoDBLocal;
