+++ doi-requests-commons/src/main/java/no/unit/nva/doi/requests/api/model/requests/CreateDoiRequest.java
@@ -1,4 +1,4 @@
-package no.unit.nva.doi.requests.api.model;
+package no.unit.nva.doi.requests.api.model.requests;
 
 import static nva.commons.utils.attempt.Try.attempt;
