+++ doi-requests-commons/src/main/java/no/unit/nva/doi/requests/service/DoiRequestsService.java
@@ -5,8 +5,8 @@ import java.net.URI;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
-import no.unit.nva.doi.requests.api.model.CreateDoiRequest;
-import no.unit.nva.doi.requests.api.model.DoiRequestSummary;
+import no.unit.nva.doi.requests.api.model.requests.CreateDoiRequest;
+import no.unit.nva.doi.requests.api.model.responses.DoiRequestSummary;
 import no.unit.nva.model.DoiRequestStatus;
 import nva.commons.exceptions.ApiGatewayException;
 import nva.commons.exceptions.ForbiddenException;
