+++ doi-requests-commons/src/main/java/no/unit/nva/doi/requests/service/impl/DynamoDBDoiRequestsService.java
@@ -27,8 +27,8 @@ import java.util.UUID;
 import java.util.stream.Collectors;
 import no.unit.nva.doi.requests.contants.ServiceConstants;
 import no.unit.nva.doi.requests.exception.DynamoDBException;
-import no.unit.nva.doi.requests.api.model.CreateDoiRequest;
-import no.unit.nva.doi.requests.api.model.DoiRequestSummary;
+import no.unit.nva.doi.requests.api.model.requests.CreateDoiRequest;
+import no.unit.nva.doi.requests.api.model.responses.DoiRequestSummary;
 import no.unit.nva.doi.requests.service.DoiRequestsService;
 import no.unit.nva.model.DoiRequest;
 import no.unit.nva.model.DoiRequestMessage;
