+++ doi-requests-commons/src/test/java/no/unit/nva/doi/requests/model/CreateDoiRequestTest.java
@@ -6,7 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import java.util.UUID;
-import no.unit.nva.doi.requests.api.model.CreateDoiRequest;
+import no.unit.nva.doi.requests.api.model.requests.CreateDoiRequest;
 import no.unit.nva.doi.requests.exception.BadRequestException;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.function.Executable;
