+++ doi-requests-commons/src/test/java/no/unit/nva/doi/requests/service/impl/DynamoDBDoiRequestsServiceTest.java
@@ -31,8 +31,8 @@ import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 import no.unit.nva.doi.requests.exception.DynamoDBException;
-import no.unit.nva.doi.requests.api.model.CreateDoiRequest;
-import no.unit.nva.doi.requests.api.model.DoiRequestSummary;
+import no.unit.nva.doi.requests.api.model.requests.CreateDoiRequest;
+import no.unit.nva.doi.requests.api.model.responses.DoiRequestSummary;
 import no.unit.nva.doi.requests.service.DoiRequestsService;
 import no.unit.nva.doi.requests.util.DoiRequestsDynamoDBLocal;
 import no.unit.nva.doi.requests.util.PublicationGenerator;
