+++ find-doi-requests/src/main/java/no/unit/nva/doi/requests/FindDoiRequestsHandler.java
@@ -15,7 +15,7 @@ import java.util.List;
 import java.util.Optional;
 import no.unit.nva.doi.requests.exception.BadRequestException;
 import no.unit.nva.doi.requests.exception.NotAuthorizedException;
-import no.unit.nva.doi.requests.api.model.DoiRequestSummary;
+import no.unit.nva.doi.requests.api.model.responses.DoiRequestSummary;
 import no.unit.nva.doi.requests.model.DoiRequestsResponse;
 import no.unit.nva.doi.requests.service.DoiRequestsService;
 import no.unit.nva.doi.requests.service.impl.DynamoDBDoiRequestsService;
